# internal/server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gemba HTTP API",
    "version": "0.1.0",
    "description": "Wire schema for the gemba server's `/api/*` surface (gm-e4.2). The SPA at `web/` consumes a typed client generated from this document; the source of truth lives in `docs/api/openapi.json` and is served at runtime under `/api/openapi.json`.\n\nThis MVP covers the read endpoints the SPA depends on day-to-day: health, version, config, repositories, work items, sprints, agents, sessions, escalations, capabilities, and the new-project conversational flow. The remaining `/api/*` surface (consults, walks, planner, perspectives) is acknowledged in `paths` with summary-only stubs and will gain full request/response schemas in follow-up beads.",
    "contact": {
      "name": "GembaCore/gemba-core",
      "url": "https://github.com/GembaCore/gemba-core"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin (the gemba server is the SPA host)."
    }
  ],
...
  "paths": {
...
    "/api/work-items": {
      "get": {
        "tags": ["work"],
        "summary": "List work items via the bound WorkPlane adaptor.",
        "operationId": "listWorkItems",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by WorkItem kind (e.g. `epic`, `task`).",
            "schema": { "type": "string" }
          },
          {
            "name": "state_category",
            "in": "query",
            "description": "Filter by normalized state category.",
            "schema": { "$ref": "#/components/schemas/StateCategory" }
          },
          {
            "name": "label",
            "in": "query",
            "description": "Filter by label (repeatable).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "WorkItem list envelope.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListWorkItemsEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/AdaptorUnavailable" }
        }
      }
    },
...
    "/api/agent-groups": {
      "get": {
        "tags": ["agents"],
        "summary": "List agent groups.",
        "operationId": "listAgentGroups",
        "responses": {
          "200": {
            "description": "Agent-group list.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListAgentGroupsEnvelope" }
              }
            }
          },
          "503": { "$ref": "#/components/responses/AdaptorUnavailable" }
        }
      }
    },
    "/api/sessions": {
      "get": {
        "tags": ["agents"],
        "summary": "List sessions known to the bound OrchestrationPlane.",
        "operationId": "listSessions",
        "responses": {
          "200": {
            "description": "Session list.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionListEnvelope" }
              }
            }
          },
          "503": { "$ref": "#/components/responses/AdaptorUnavailable" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Malformed request — invalid query parameters or body.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "AdaptorUnavailable": {
        "description": "Adaptor is not registered or is degraded; the SPA shows a banner and keeps cached data visible.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Shared error envelope (gm-faz). `error` is a stable token the SPA branches on; `message` is human-readable and may be surfaced verbatim.",
        "required": ["error", "message"],
        "properties": {
          "error": { "type": "string", "examples": ["adaptor_degraded", "validation", "not_found"] },
          "message": { "type": "string" }
        }
      },
...
      "StateCategory": {
        "type": "string",
        "enum": ["backlog", "unstarted", "staged", "started", "completed", "canceled"]
      },
      "AgentRef": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" }
        }
      },
      "WorkItem": {
        "type": "object",
        "description": "Subset of the WorkItem fields most commonly read by the SPA. The full shape (relationships, evidence, dod, derived signals, custom fields) is the source of truth in `core/types.go` and `web/src/types/core.gen.ts`; this schema documents the wire-stable subset for a typed client. `additionalProperties: true` keeps clients forward-compatible with fields landing in follow-up beads.",
        "additionalProperties": true,
        "required": ["id", "kind", "title", "status", "state_category", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string" },
          "kind": { "type": "string", "examples": ["epic", "task", "feature", "decision"] },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "state_category": { "$ref": "#/components/schemas/StateCategory" },
          "priority": { "type": ["integer", "null"], "minimum": 0, "maximum": 4 },
          "owner": { "$ref": "#/components/schemas/AgentRef" },
          "assignee": { "$ref": "#/components/schemas/AgentRef" },
          "labels": { "type": "array", "items": { "type": "string" } },
          "primary_repository_id": { "type": "string" },
          "repository_ids": { "type": "array", "items": { "type": "string" } },
          "sprint_id": { "type": ["string", "null"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ListWorkItemsEnvelope": {
        "type": "object",
        "required": ["items", "total"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WorkItem" }
          },
          "total": { "type": "integer" }
        }
      },
...
      "AgentGroup": {
        "type": "object",
        "additionalProperties": true,
        "required": ["id"],
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "members": { "type": "array", "items": { "type": "string" } }
        }
      },
      "ListAgentGroupsEnvelope": {
        "type": "object",
        "required": ["agent_groups"],
        "properties": {
          "agent_groups": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AgentGroup" }
          }
        }
      },
      "Session": {
        "type": "object",
        "additionalProperties": true,
        "required": ["id"],
        "properties": {
          "id": { "type": "string" },
          "agent_id": { "type": "string" },
          "status": { "type": "string" },
          "started_at": { "type": "string", "format": "date-time" }
        }
      },
      "SessionListEnvelope": {
        "type": "object",
        "required": ["sessions"],
        "properties": {
          "sessions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Session" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token auth (gm-e5.1). Required when the server boots with `--auth=token`."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gemba_session",
        "description": "Signed session cookie issued by `POST /api/auth/login` after a bearer exchange."
      }
    }
  },
  "security": [
    { "bearerToken": [] },
    { "sessionCookie": [] },
    {}
  ]
}